Relocate Indemnity allocation Calculation method
